Change logic to use 2nd oscillatory cycle only.
Instead of checking both positive and negative
transitions to get half the period, I now check
only two consecutive negative input transitions,
and in that only the 2nd and 3rd negative
transition. This way, I get the period of the 2nd
cycle of the LC tank, which is much less prone to
input noises.